requirements.txt
langchain
langchain_community
langchain_openai
typing
operator
langchain_core
pandas
streamlit
python-dotenv
Chroma
pydantic
openai